DataStore 0.0.2
The datastore is now relying on JSON instead of DBObject for its serialization/deserialization public interface.

+++ net.modelbased.sensapp.datastore/src/test/scala/net/modelbased/sensapp/datastore/specs/data/DataModelRegistry.scala
@@ -37,13 +37,13 @@ sealed abstract class DataModel(val n: String)
  * MultiTypedData illustrates a persistent object made with a String and an Int.
  * @author Sebastien Mosser
  */
-case class MultiTypedData(x: String, v: Int) extends DataModel(x)
+case class MultiTypedData(x: String, v: Long) extends DataModel(x)
 
 /**
  * SequenceData illustrates how persistent collections of scalar data are handled
  * @author Sebastien Mosser
  */
-case class SequenceData(x: String, v: List[Int]) extends DataModel(x)
+case class SequenceData(x: String, v: List[Long]) extends DataModel(x)
 
 /**
  * TypedSequenceData illustrates how persistent collections of complex objects are handled
